requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.5
pycparser==2.22
pycryptodome==3.23.0
python-dotenv==1.1.1
